requirements.txt
pytest_bdd
pytest~=8.3.4
selenium~=3.141.0
allure-pytest~=2.13.5
PyMySQL~=1.1.1
pony~=0.7.19
jsonpickle~=4.0.1
pytest-bdd~=8.1.0
allure-python-commons~=2.13.5